+++ ts_runtime/src/tun_actor.rs
@@ -366,6 +366,7 @@ mod tests {
             underlay_addresses: vec![],
             derp_region: None,
             cap: Default::default(),
+            cap_map: Default::default(),
             peerapi_port: None,
             peerapi_dns_proxy: false,
             is_wireguard_only: false,